Fix more invalid mocks pointing to old paths and one more renaming
