+++ src/databricks/labs/lakebridge/reconcile/trigger_recon_aggregate_service.py
@@ -16,7 +16,7 @@ from databricks.labs.lakebridge.reconcile.recon_output_config import (
     AggregateQueryOutput,
     DataReconcileOutput,
 )
-from databricks.labs.lakebridge.reconcile.reconcilation import Reconciliation
+from databricks.labs.lakebridge.reconcile.reconciliation import Reconciliation
 from databricks.labs.lakebridge.reconcile.trigger_recon_service import TriggerReconService
 from databricks.labs.lakebridge.reconcile.schema_service import SchemaService
 from databricks.labs.lakebridge.reconcile.table_service import NormalizeReconConfigService
