+++ tests/integration/reconcile/test_aggregates_reconcile.py
@@ -10,7 +10,7 @@ from pyspark.testing import assertDataFrameEqual
 from pyspark.sql import Row
 
 from databricks.labs.lakebridge.config import DatabaseConfig, ReconcileMetadataConfig
-from databricks.labs.lakebridge.reconcile.reconcilation import Reconciliation
+from databricks.labs.lakebridge.reconcile.reconciliation import Reconciliation
 from databricks.labs.lakebridge.transpiler.sqlglot.dialect_utils import get_dialect
 from databricks.labs.lakebridge.reconcile.connectors.data_source import MockDataSource
 from databricks.labs.lakebridge.reconcile.execute import main
